UserDto, ClusterDto, UserService, ClusterService refactor

+++ src/main/java/com/example/seven/service/ClusterService.java
@@ -9,6 +9,9 @@ import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Isolation;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 @Service
@@ -18,22 +21,30 @@ public class ClusterService {
     private final ClusterRepository clusterRepository;
     private final CloudService cloudService;
 
-    public Page<Cluster> findAll(Pageable pageable, String username) {
-        return clusterRepository.findAllByUserUsername(pageable, username);
+    public Page<ClusterDto> findAllByUserUsername(Pageable pageable, String username) {
+        return clusterRepository.findAllByUserUsername(pageable, username)
+                .map(ClusterMapper.CLUSTER_MAPPER::clusterToClusterDto);
     }
 
-    public Cluster findOne(Long id) {
-        return clusterRepository.findById(id).orElse(null);
+    public ClusterDto findOne(Long id) {
+        Cluster cluster = clusterRepository.findById(id).orElse(null);
+        if (cluster != null) {
+            return ClusterMapper.CLUSTER_MAPPER.clusterToClusterDto(cluster);
+        }
+        return null;
     }
 
-    public Cluster create(User user, ClusterDto clusterDto) {
+    @Transactional(rollbackFor = Exception.class, isolation = Isolation.DEFAULT, propagation = Propagation.REQUIRED)
+    public ClusterDto create(User user, ClusterDto clusterDto) {
         Cluster cluster = ClusterMapper.CLUSTER_MAPPER.clusterDtoToCluster(clusterDto);
         cluster.setUser(user);
-        return clusterRepository.save(cluster);
+        clusterRepository.save(cluster);
+        return ClusterMapper.CLUSTER_MAPPER.clusterToClusterDto(cluster);
     }
 
+    @Transactional(rollbackFor = Exception.class, isolation = Isolation.DEFAULT, propagation = Propagation.REQUIRED)
     public String uploadImg(Long id, MultipartFile file) {
-        Cluster cluster = findOne(id);
+        Cluster cluster = clusterRepository.findById(id).orElse(null);
         if (cluster != null && file != null) {
             String url = cloudService.getUrl(file);
             cluster.setImgUrl(url);
@@ -43,25 +54,35 @@ public class ClusterService {
         return null;
     }
 
-    public Cluster update(Long id, ClusterDto clusterDto) {
-        Cluster cluster = findOne(id);
+    @Transactional(rollbackFor = Exception.class, isolation = Isolation.DEFAULT, propagation = Propagation.REQUIRED)
+    public ClusterDto update(Long id, ClusterDto clusterDto) {
+        Cluster cluster = clusterRepository.findById(id).orElse(null);
         if (cluster != null) {
             cluster.setName(clusterDto.getName());
             cluster.setDescription(clusterDto.getDescription());
             cluster.setFieldOne(clusterDto.getFieldOne());
             cluster.setFieldTwo(clusterDto.getFieldTwo());
             cluster.setFieldThree(clusterDto.getFieldThree());
-            return clusterRepository.save(cluster);
+            clusterRepository.save(cluster);
+            return ClusterMapper.CLUSTER_MAPPER.clusterToClusterDto(cluster);
         }
         return null;
     }
 
-    public void delete(Long id){
-        Cluster cluster = findOne(id);
-        clusterRepository.delete(cluster);
+    @Transactional(rollbackFor = Exception.class, isolation = Isolation.DEFAULT, propagation = Propagation.REQUIRED)
+    /*need optimize*/
+    public void delete(Long id) {
+        Cluster cluster = clusterRepository.findById(id).orElse(null);
+        if (cluster != null) {
+            cluster.getItems().forEach(item ->
+                    item.getTags().forEach(tag ->
+                            tag.getItems().remove(item)));
+            clusterRepository.delete(cluster);
+        }
     }
 
-    public Page<Cluster> findAllOrderByItemsSize(Pageable pageable){
-        return clusterRepository.findAllOrderByItemsSize(pageable);
+    public Page<ClusterDto> findAllOrderByItemsSize(Pageable pageable) {
+        return clusterRepository.findAllOrderByItemsSize(pageable)
+                .map(ClusterMapper.CLUSTER_MAPPER::clusterToClusterDto);
     }
 }
